typing: clarify types in `config.py`, `module.py`
`repo_path` can be both a `str` and a `Path`, so update the
type to a `Union` of the two. `validate` returns `List[str]`
but it can also return `None` if there are no validation errors,
so mark it as `Optional`.

Also remove an unused instance of `i` and the `enumerate` that
causes it.